Extracts the call to the presence handler from switch-case destined for requests (#39)
In onMessage, that occurs when the MQTT client receives a message, there is a
switch protected by a mutex for each case of a call to c.request that results in
a "dispatch" of the response. However, Presence doesn't use that pattern and has
therefore no reason to be inside that switch-case.